making Satellite layer rendering more efficient
closes #46

I tried changing the actual base map, but that caused the hex layer to have to reload and render again. The fix here is I just added all the layers we want for the normal map to an array and layers we want visible for the satellite map to another array. I show/hide the appropriate ones based on user selection.